[node][memory] Garbage collect cache sooner
In nodejs we are able to garbage collect memory quicker if we don't
keep a reference to it from the cleanup timeout.

This is particularly important to services with short-lived throttled
functions such as our mail worker.

Since browsers don't support native modules and don't provide
[WeakRef](https://github.com/tc39/proposal-weakrefs) yet we keep the
old behaviour. It should also be less problematic there since throttled
functions are not being created in the same amounts.